peluqueria_backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
pyodbc==4.0.32
python-dotenv==0.19.0